refactor(python): inline apply_dry_run_to_request into crd.py; drop mutation_options module
